+++ src/layers/controllers/post.controller.js
@@ -84,7 +84,6 @@ const findOnePost = async (req, res, next) => {
 
 /** @param { e.Request } req @param { e.Response } res @param { e.NextFunction } next */
 const findHotPosts = async (req, res, next) => {
-    // await postService.createHotPosts();
     const hotPosts = await postService.findHotPosts();
 
     return res.status(200).json(new exception.FormDto('인기 게시물 조회 성공', { hotPosts }));
