+++ startMsg.js
@@ -1,6 +1,7 @@
 const { Post } = require('./src/sequelize/models');
 const sequelize = require('sequelize');
-
+const scheduller = require('./src/layers/exceptModels/form/scheduller');
+scheduller.createHotPost;
 const Op = sequelize.Op;
 // const fs = require('fs');
 
@@ -51,16 +52,16 @@ const Op = sequelize.Op;
 // status = !status;
 // console.log(status);
 
-const data = [
-    { name: '권영', createdAt: 0911 },
-    { name: '수민', createdAt: 0910 },
-    { name: '수범', createdAt: 0909 }
-];
-const calcurateCreatedAt = (createdAt) => {
-    return createdAt + 10;
-};
+// const data = [
+//     { name: '권영', createdAt: 0911 },
+//     { name: '수민', createdAt: 0910 },
+//     { name: '수범', createdAt: 0909 }
+// ];
+// const calcurateCreatedAt = (createdAt) => {
+//     return createdAt + 10;
+// };
 
-for (let gwonyeong of data) {
-    gwonyeong.createdAt = calcurateCreatedAt(At.createdAt);
-}
-console.log(data);
+// for (let gwonyeong of data) {
+//     gwonyeong.createdAt = calcurateCreatedAt(At.createdAt);
+// }
+// console.log(data);
